improve imports and rename PartialMessagesError to PartialError
